Update `run-ngs-pipeline`
* Update :command:`run-ngs-pipeline` command to check whether input gene 
is one of the target genes before attempting to run the pipeline.

+++ pypgx/api/pipeline.py
@@ -41,6 +41,9 @@ def run_ngs_pipeline(
     do_not_plot_allele_fraction : bool, default: False
         Do not plot allele fraction profile.
     """
+    if not utils.is_target_gene(gene):
+        raise utils.NotTargetGeneError(gene)
+
     if os.path.exists(output) and force:
         shutil.rmtree(output)
